Add display file content in the form of a table
When opening a file, its contents are displayed as a table with 17 columns. The first column represents the offset of the byte position. The rest contain byte values. When you click Close button, the content disappears. The display function works slow. It is due to the features of the FileTableModel, so it needs to be fixed.